app/requirements.txt
aioredis
async-timeout
attrs
cleo
click
crashtest
Deprecated
exceptiongroup
Flask[async]
SQLAlchemy
Alembic
psycopg2-binary
geographiclib
geopy
gunicorn
importlib-metadata
iniconfig
itsdangerous
Jinja2
MarkupSafe
more-itertools
mypy
numpy
packaging
pluggy
pptree
pylev
pyparsing
python-dotenv
python-ulid
scipy
tomli
Werkzeug
wrapt
zipp
